data_corpus_amicus {quanteda.corpora}	R Documentation

Amicus curiae briefs from Bakke (1978) and Bollinger (2008)

Description

Texts of amicus curiae briefs labelled as being either pro-petitioner or pro-respondent in US Supreme court cases on affirmative action, Bakke (1978) and Bollinger (2008), taken from Evans et al (2007).

Usage

data_corpus_amicus

Format

amicusCorpus is corpus with 100 texts, and docvar set for petitioner/respondent

References

Evans, Michael, et al. 2007. "Recounting the courts? Applying automated content analysis to enhance empirical legal research." Journal of Empirical Legal Studies 4(4): 1007-1039.

[Package quanteda.corpora version 0.85 Index]