data_corpus_immigrationnews {quanteda.corpora}	R Documentation

UK news articles (2,833) from 2014 that mention immigration

Description

A corpus of articles form the UK press in 2014 that mention the immigration. See working paper: "The Immigration Issue in the European Electoral Campaign in the UK: Text-Mining Public Debate from Newspapers and Social Media"

Usage

data_corpus_immigrationnews

Format

immigNewsCorpus is corpus with 2833 texts, and docvars set for paperName, day, and id. day is the number of days from Jan 1st 2014.

[Package quanteda.corpora version 0.85 Index]