data_corpus_irishbudgets {quanteda.corpora}	R Documentation

Annual budget speeches from the Irish Dail, 2008-2012

Description

Speeches and document-level variables from Irish budget debates held annually, for the years 2008-2012.

Usage

data_corpus_irishbudgets

Format

year 4-digit year of the debate; debate character indicating whether the debate concerned a main debate ("BUDGET" or supplementary budget "BUDGETSUP"); number two-digit number as a character, indicating the order of the speech within the debate; namefirst and namelast first and last name of the speaker; and party a character abbrevation of the political party of the speaker.

References

Alexander Herzog and Kenneth Benoit. 2015. "The Most Unkindest Cuts: Speaker Selection and Expressed Government Dissent During Economic Crisis". Journal of Politics 77(4, October): 1157-1175.

Lowe, Will and Kenneth Benoit. 2013. "Validating Estimates of Latent Traits From Textual Data Using Human Judgment as a Benchmark." Political Analysis 21: 298-313.

[Package quanteda.corpora version 0.85 Index]