data_corpus_sotu {quanteda.corpora}	R Documentation

U.S. State of the Union addresses from 1790 to 2016

Description

A corpus object of every US State-of-the-Union address, from 1790 to present.

Usage

data_corpus_sotu

Format

The corpus includes the following document variables:

FirstName

President's first and middle names.

President

President's last name.

Date

Date of the delivery of the speech or document.

delivery

Either written or spoken, depending on the format. See Source.

type

Either SOTU for an official State of the Union Address, or other, for a different form of speech or message. See Source.

party

President's party.

Source

The American Presidency Project, http://www.presidency.ucsb.edu/sou.php.

[Package quanteda.corpora version 0.85 Index]