download {quanteda.corpora}	R Documentation

Download an RDS file from online locations

Description

Download an RDS file from online locations

Usage

download(name = NULL, url = NULL, cache = TRUE, ...)

Arguments

name
name of a RDS file to download

url
user-specified location of the RDS file to download

cache
if TRUE, save downloaded file to temporary folder

...
extra arguments passed to download.file

Details

Datasets currently available by name are:

data_corpus_guardian

6,000 Guradian newspaper articles in politics, economy, society and international sections from 2012 to 2016.

data_corpus_foreignaffairscommittee

Transcripts of all the speeches at Japan's Committee on Foreign Affairs and Defense of the lower house (Shugiin) from 1947 to 2017

.

Value

an unserialized R object

Examples

## Not run: 
corp <- download('data_corpus_guardian')
corp <- download('data_corpus_foreignaffairscommittee')

## End(Not run)

[Package quanteda.corpora version 0.85 Index]