quanteda.corpora-package {quanteda.corpora}	R Documentation

An R package containing texts and corpora for quanteda.

Description

A set of texts and corpus objects for use with the quanteda R package.

Author(s)

Maintainer: Kenneth Benoit kbenoit@lse.ac.uk

Other contributors:

Kohei Watanabe watanabe.kohei@gmail.com [contributor]

See Also

quanteda

[Package quanteda.corpora version 0.85 Index]